insight_whisper/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
yt-dlp==2025.3.31
openai-whisper==20240930
google-cloud-storage==2.19.0
google-cloud-pubsub==2.26.1
